+++ src/components/utils/get-window.ts
@@ -1,7 +1,7 @@
 // For the purposes of correctly mocking the window object in Jest,
 // this is a safe way of returning it for tests to work as expected.
 export default function getWindow() {
-  if (typeof window === undefined) {
+  if (!window) {
     throw new Error('window not available');
   }